requirements.txt
numpy 2.0.1
opencv-python~=4.10.0.84
python-dotenv~=1.0.1